fix(DTXReactNativeSupport): JS and Content can load in any order (#74)
Previously waitForReactNativeLoadWithCompletionHandler would
add an observer for "RCTJavaScriptDidLoadNotification",
then when notified would add a second observer for
"RCTContentDidAppearNotification", and after getting a notification
for both it would call the completion handler.

However, with the React Native new architecture the loading of
the JS and Content can happen in a different order (content first).
This would cause DetoxSync to get stuck when enabling synchronization.

Instead, we register all observers (JS, content, failure) one after
the other, and we use a race-free C11 atomic to count when we
receive a successful JS or Content notification.
After we receive both (or fail), we call the handler as before.

Co-authored-by: tux3 <barrdetwix@gmail.com>
Co-authored-by: Asaf Korem <55082339+asafkorem@users.noreply.github.com>